Move "routesDisabled" check into private session middleware
This way both HTTP and non-HTTP requests can enjoy the benefits of `routesDisabled`.  So many benefits.
