+++ lib/hooks/http/get-configured-http-middleware-fns.js
@@ -79,29 +79,6 @@ module.exports = function getBuiltInHttpMiddleware (expressRouterMiddleware, sai
 
         return function (req, res, next){
 
-          // If configured to do so (i.e. there is at least one entry in the `sails.hooks.session.routesDisabled` blacklist)
-          // then check this request against each entry in the blacklist and skip running session middleware if this is a match.
-          var isSessionDisabled = _.any(sails.hooks.session.routesDisabled, function (disabledRouteInfo){
-
-            // Figure out if the request's method matches.
-            var isMethodExactMatch = req.method === disabledRouteInfo.method;
-            var isMethodImplicitMatch = disabledRouteInfo.method === 'ALL' || (disabledRouteInfo.method === '' && _.contains(['GET', 'POST', 'PUT', 'DELETE', 'PATCH'], req.method));
-            // If not, then skip this disabled route- it's not a match.
-            if (!isMethodExactMatch && !isMethodImplicitMatch) {
-              return;
-            }
-
-            // Then figure out if the request's url path matches.
-            var isUrlPathMatch = req.path.match(disabledRouteInfo.urlPatternRegExp);
-            return isUrlPathMatch;
-
-          });//</_.any()>
-
-          // If the session is disabled, then skip running the middleware.
-          if (isSessionDisabled) {
-            return next();
-          }
-
           // --•
           // Run the session middleware.
           configuredSessionMiddleware(req,res,function (err) {
