+++ test/unit/req.session.test.js
@@ -52,59 +52,89 @@ describe('req.session', function (){
         res.send();
       });
 
-    });
-
-    it('should exist', function (done) {
-      app.request({
-        url: '/sessionTest',
-        method: 'POST',
-        params: {},
-        headers: {}
-      }, function (err, res, body){
-        if (err) return done(err);
-        if (res.statusCode !== 200) return done(new Error('Expected 200 status code'));
-        if (!doesSessionExist) return done(new Error('req.session should exist.'));
-        if (!isSessionAnObject) return done(new Error('req.session should be an object.'));
-        return done();
+      app.get('/sails.io.js', function (req, res){
+        doesSessionExist = !!req.session;
+        res.send();
       });
+
     });
 
+    describe('with routes in the routesDisabled list', function() {
 
-    //
-    // To test:
-    //
-    // DEBUG=express-session mocha test/unit/req.session.test.js -b -g 'should persist'
-    //
-
-    it('should persist data between requests', function (done){
-      app.request({
-        url: '/sessionTest',
-        method: 'POST',
-        params: {},
-        headers: {}
-      }, function (err, clientRes, body){
-        if (err) return done(err);
-        if (clientRes.statusCode !== 200) return done(new Error('Expected 200 status code'));
-        if (!doesSessionExist) return done(new Error('req.session should exist.'));
-        if (!isSessionAnObject) return done(new Error('req.session should be an object.'));
-
+      it('should not exist', function (done) {
         app.request({
-          url: '/sessionTest',
+          url: '/sails.io.js',
           method: 'GET',
           params: {},
-          headers: {
-            cookie: clientRes.headers['set-cookie']
-          }
+          headers: {}
+        }, function (err, res, body){
+          if (err) return done(err);
+          if (res.statusCode !== 200) return done(new Error('Expected 200 status code'));
+          if (doesSessionExist) return done(new Error('req.session should not exist.'));
+          if (res.headers['set-cookie']) return done(new Error('Should not have a `set-cookie` header in the response.'));
+          return done();
+        });
+      });
+
+    });
+
+    describe('with routes NOT in the routesDisabled list', function() {
+
+      it('should exist', function (done) {
+        app.request({
+          url: '/sessionTest',
+          method: 'POST',
+          params: {},
+          headers: {}
+        }, function (err, res, body){
+          if (err) return done(err);
+          if (res.statusCode !== 200) return done(new Error('Expected 200 status code'));
+          if (!doesSessionExist) return done(new Error('req.session should exist.'));
+          if (!isSessionAnObject) return done(new Error('req.session should be an object.'));
+          return done();
+        });
+      });
+
+
+      //
+      // To test:
+      //
+      // DEBUG=express-session mocha test/unit/req.session.test.js -b -g 'should persist'
+      //
+
+      it('should persist data between requests', function (done){
+        app.request({
+          url: '/sessionTest',
+          method: 'POST',
+          params: {},
+          headers: {}
         }, function (err, clientRes, body){
           if (err) return done(err);
           if (clientRes.statusCode !== 200) return done(new Error('Expected 200 status code'));
           if (!doesSessionExist) return done(new Error('req.session should exist.'));
           if (!isSessionAnObject) return done(new Error('req.session should be an object.'));
-          if (!doesTestPropertyStillExist) return done(new Error('`req.session.something` should still exist for subsequent requests.'));
-          return done();
+
+          app.request({
+            url: '/sessionTest',
+            method: 'GET',
+            params: {},
+            headers: {
+              cookie: clientRes.headers['set-cookie']
+            }
+          }, function (err, clientRes, body){
+            if (err) return done(err);
+            if (clientRes.statusCode !== 200) return done(new Error('Expected 200 status code'));
+            if (!doesSessionExist) return done(new Error('req.session should exist.'));
+            if (!isSessionAnObject) return done(new Error('req.session should be an object.'));
+            if (!doesTestPropertyStillExist) return done(new Error('`req.session.something` should still exist for subsequent requests.'));
+            return done();
+          });
         });
       });
+
     });
+
+
 
   });
